feat: add lan chat in programs folder
there is a markdown file containing notes, sort of like a readme which explains the usage of lan chat. there are 2 types of lan chat, ecdh and mini. mini lan chat does not include any encryption but is much simpler. ecdh lan chat uses elliptic curve diffie-hellman encryption, so clients have end-to-end encryption with the server. each of the types of lan chat contain a server and a client. the server displays all messages while the client sends and receives messages. they feature parallel threads so messages can be send and recieve. for more detail, see the notes file